% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_template_path}
\alias{bootstrap_template_path}
\title{Path to the installed bootstrap template}
\usage{
bootstrap_template_path()
}
\value{
Character path or \code{NULL}.
}
\description{
Path to the installed bootstrap template
}
\keyword{internal}
