% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discover_bindings_js}
\alias{discover_bindings_js}
\title{Read the UI-discovery JS script as a single string}
\usage{
discover_bindings_js()
}
\value{
Character scalar.
}
\description{
Read the UI-discovery JS script as a single string
}
\keyword{internal}
