% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{read_id_srcrefs}
\alias{read_id_srcrefs}
\title{Read the id -> srcref attribution RDS file, if present}
\usage{
read_id_srcrefs(app_dir)
}
\arguments{
\item{app_dir}{Path to the Shiny app directory.}
}
\value{
Named list (id -> srcref), possibly empty.
}
\description{
Read the id -> srcref attribution RDS file, if present
}
\keyword{internal}
