% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rprofile_backup_path}
\alias{rprofile_backup_path}
\title{Path to the .Rprofile backup file}
\usage{
rprofile_backup_path(app_dir)
}
\arguments{
\item{app_dir}{Path to the Shiny app directory.}
}
\value{
Character path.
}
\description{
Path to the .Rprofile backup file
}
\keyword{internal}
