% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{shinycov_get_counters}
\alias{shinycov_get_counters}
\title{Get the currently-active counters environment}
\usage{
shinycov_get_counters()
}
\value{
An environment.
}
\description{
Get the currently-active counters environment
}
\keyword{internal}
