% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinycov_input_actions}
\alias{shinycov_input_actions}
\title{Action names that count as interacting with an input}
\usage{
shinycov_input_actions()
}
\value{
Character vector.
}
\description{
Action names that count as interacting with an input
}
\keyword{internal}
