% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinycov_output_actions}
\alias{shinycov_output_actions}
\title{Action names that count as verifying an output}
\usage{
shinycov_output_actions()
}
\value{
Character vector.
}
\description{
Action names that count as verifying an output
}
\keyword{internal}
