% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinycov_output_file}
\alias{shinycov_output_file}
\title{Full path to the coverage RDS file}
\usage{
shinycov_output_file(app_dir)
}
\arguments{
\item{app_dir}{Path to the Shiny app directory.}
}
\value{
Character path to \code{coverage.rds}.
}
\description{
Full path to the coverage RDS file
}
\keyword{internal}
