% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinycov_bootstrap_path}
\alias{shinycov_bootstrap_path}
\title{Full path to the bootstrap R script}
\usage{
shinycov_bootstrap_path(app_dir)
}
\arguments{
\item{app_dir}{Path to the Shiny app directory.}
}
\value{
Character path to `bootstrap.R`.
}
\description{
Full path to the bootstrap R script
}
\keyword{internal}
