% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinycov_output_dir}
\alias{shinycov_output_dir}
\title{Determine the shiny.cov output directory for a given app directory}
\usage{
shinycov_output_dir(app_dir)
}
\arguments{
\item{app_dir}{Path to the Shiny app directory.}
}
\value{
Character path to the `.shiny.cov/` subdirectory.
}
\description{
Determine the shiny.cov output directory for a given app directory
}
\keyword{internal}
